move notes table to shared folder
