+++ apps/game-master-client/app/routes/_app.games.$gameId.notes._index/notes-index.tsx
@@ -1,7 +1,7 @@
 import { useTypedLoaderData } from "remix-typedjson";
 import { CreateNoteSlideover } from "~/components/forms/create-note";
-import { NoteTable } from "./components/note-table";
 import type { loader } from "./route";
+import { NoteTable } from "~/components/tables/note-table";
 
 export function NotesIndex() {
   const { allGameNotes, gameId } = useTypedLoaderData<typeof loader>();
